+++ include/CLI/CLI.hpp
@@ -24,6 +24,12 @@
 
 #include "CLI/Validators.hpp"
 
+#include "CLI/Formatter.hpp"
+
 #include "CLI/Option.hpp"
 
+#include "CLI/OptionFormatter.hpp"
+
 #include "CLI/App.hpp"
+
+#include "CLI/AppFormatter.hpp"